+++ LoginModul_Test/UserChangeEMail_Test.cs
@@ -15,7 +15,7 @@ namespace LoginModul_Test
                 LoginModul.LoginModul lm = new LoginModul.LoginModul(new FakeLoginDataMapper(),null, new LoginModul.EMailCheck(FakeHelper.chkEMail), new LoginModul.PasswordCheck(FakeHelper.chkPassword), FakeHelper.FakeErrorMessage);
                 string verificationCode = lm.CreateLogin("bejo@eal.dk", "123456", "123456");
                 lm.ActivateUser("bejo@eal.dk", verificationCode);
-                int userID = lm.Login("bejo@eal.dk", "123456");
+                long userID = lm.Login("bejo@eal.dk", "123456");
                 //lm.ChangeEmail("bejo@eal.dk", "bbb@eal.dk", "123456");
                 DidItThoughAnException = false;
             }
