+++ LoginModul_Test/UserResetPassword_Test.cs
@@ -15,12 +15,12 @@ namespace LoginModul_Test
                 LoginModul.LoginModul lm = new LoginModul.LoginModul(new FakeLoginDataMapper(),null, new LoginModul.EMailCheck(FakeHelper.chkEMail), new LoginModul.PasswordCheck(FakeHelper.chkPassword), FakeHelper.FakeErrorMessage);
                 string verificationCode = lm.CreateLogin("bejo@eal.dk", "123456", "123456");
                 lm.ActivateUser("bejo@eal.dk", verificationCode);
-                int userID1 = lm.Login("bejo@eal.dk", "123456"); // Poof it works
+                long userID1 = lm.Login("bejo@eal.dk", "123456"); // Poof it works
 
                 string newVerificationCode = lm.InitiateResetPassword("bejo@eal.dk");
                 lm.ResetPassword("bejo@eal.dk", newVerificationCode, "NewPassword", "NewPassword");
 
-                int userID2 = lm.Login("bejo@eal.dk", "NewPassword"); // Poof it works again
+                long userID2 = lm.Login("bejo@eal.dk", "NewPassword"); // Poof it works again
 
                 DidItThoughAnException = (userID1 != userID2);
             }
@@ -40,11 +40,11 @@ namespace LoginModul_Test
                 LoginModul.LoginModul lm = new LoginModul.LoginModul(new FakeLoginDataMapper(),null, new LoginModul.EMailCheck(FakeHelper.chkEMail), new LoginModul.PasswordCheck(FakeHelper.chkPassword), FakeHelper.FakeErrorMessage);
                 string verificationCode = lm.CreateLogin("bejo@eal.dk", "123456", "123456");
                 lm.ActivateUser("bejo@eal.dk", verificationCode);
-                int userID1 = lm.Login("bejo@eal.dk", "123456"); // Poof it works
+                long userID1 = lm.Login("bejo@eal.dk", "123456"); // Poof it works
 
                 string newVerificationCode = lm.InitiateResetPassword("bejo@eal.dk");
 
-                int userID2 = lm.Login("bejo@eal.dk", "123456"); // Poof it works again
+                long userID2 = lm.Login("bejo@eal.dk", "123456"); // Poof it works again
 
                 DidItThoughAnException = (userID1 != userID2);
             }
